django_json/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
data-wizard==1.3.0
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
html-json-forms==1.1.1
idna==2.9
itertable==2.0.0
Markdown==3.2.2
natural-keys==1.6.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
xlrd==1.2.0
